requirements.txt
Python Libraries needed:
 - pandas
 - matplotlib
 - numpy
 - scipy
 - datetime
 - math
 - typing

 
